Move resolve-user-path to utils dir
