+++ packages/react-cosmos-webpack/src/utils/resolve-user-path.js
